Inline AssistanceActionService into AssistanceController
The service was legacy code that took Database.Connection instead of
Transaction and didn't autowire any beans. All methods are now inlined
directly into the controller, using proper transaction blocks.
