+++ service/src/main/kotlin/evaka/core/assistance/AssistanceController.kt
@@ -10,8 +10,12 @@ import evaka.core.assistanceaction.AssistanceAction
 import evaka.core.assistanceaction.AssistanceActionOption
 import evaka.core.assistanceaction.AssistanceActionRequest
 import evaka.core.assistanceaction.AssistanceActionResponse
-import evaka.core.assistanceaction.AssistanceActionService
+import evaka.core.assistanceaction.deleteAssistanceAction
+import evaka.core.assistanceaction.getAssistanceActionOptions
 import evaka.core.assistanceaction.getAssistanceActionsByChild
+import evaka.core.assistanceaction.insertAssistanceAction
+import evaka.core.assistanceaction.shortenOverlappingAssistanceAction
+import evaka.core.assistanceaction.updateAssistanceAction
 import evaka.core.shared.AssistanceActionId
 import evaka.core.shared.AssistanceFactorId
 import evaka.core.shared.ChildId
@@ -23,9 +27,12 @@ import evaka.core.shared.async.AsyncJobRunner
 import evaka.core.shared.auth.AuthenticatedUser
 import evaka.core.shared.data.DateSet
 import evaka.core.shared.db.Database
+import evaka.core.shared.db.mapPSQLException
+import evaka.core.shared.domain.BadRequest
 import evaka.core.shared.domain.EvakaClock
 import evaka.core.shared.security.AccessControl
 import evaka.core.shared.security.Action
+import org.jdbi.v3.core.JdbiException
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
@@ -37,7 +44,6 @@ import org.springframework.web.bind.annotation.RestController
 @RestController
 class AssistanceController(
     private val accessControl: AccessControl,
-    private val assistanceActionService: AssistanceActionService,
     private val asyncJobRunner: AsyncJobRunner<AsyncJob>,
 ) {
     data class AssistanceFactorResponse(
@@ -169,22 +175,27 @@ class AssistanceController(
         @RequestBody body: AssistanceActionRequest,
     ): AssistanceAction {
         return db.connect { dbc ->
-                dbc.read {
+                dbc.transaction { tx ->
                     accessControl.requirePermissionFor(
-                        it,
+                        tx,
                         user,
                         clock,
                         Action.Child.CREATE_ASSISTANCE_ACTION,
                         childId,
                     )
+                    try {
+                        validateActions(body, tx.getAssistanceActionOptions())
+                        tx.shortenOverlappingAssistanceAction(
+                            user,
+                            clock.now(),
+                            childId,
+                            body.startDate,
+                        )
+                        tx.insertAssistanceAction(user, clock.now(), childId, body)
+                    } catch (e: JdbiException) {
+                        throw mapPSQLException(e)
+                    }
                 }
-                assistanceActionService.createAssistanceAction(
-                    dbc,
-                    user = user,
-                    now = clock.now(),
-                    childId = childId,
-                    data = body,
-                )
             }
             .also { assistanceAction ->
                 Audit.ChildAssistanceActionCreate.log(
@@ -203,22 +214,21 @@ class AssistanceController(
         @RequestBody body: AssistanceActionRequest,
     ): AssistanceAction {
         return db.connect { dbc ->
-                dbc.read {
+                dbc.transaction { tx ->
                     accessControl.requirePermissionFor(
-                        it,
+                        tx,
                         user,
                         clock,
                         Action.AssistanceAction.UPDATE,
                         id,
                     )
+                    try {
+                        validateActions(body, tx.getAssistanceActionOptions())
+                        tx.updateAssistanceAction(user, clock.now(), id, body)
+                    } catch (e: JdbiException) {
+                        throw mapPSQLException(e)
+                    }
                 }
-                assistanceActionService.updateAssistanceAction(
-                    dbc,
-                    user = user,
-                    now = clock.now(),
-                    id = id,
-                    data = body,
-                )
             }
             .also { Audit.ChildAssistanceActionUpdate.log(targetId = AuditId(id)) }
     }
@@ -231,16 +241,16 @@ class AssistanceController(
         @PathVariable id: AssistanceActionId,
     ) {
         db.connect { dbc ->
-            dbc.read {
+            dbc.transaction { tx ->
                 accessControl.requirePermissionFor(
-                    it,
+                    tx,
                     user,
                     clock,
                     Action.AssistanceAction.DELETE,
                     id,
                 )
+                tx.deleteAssistanceAction(id)
             }
-            assistanceActionService.deleteAssistanceAction(dbc, id)
         }
         Audit.ChildAssistanceActionDelete.log(targetId = AuditId(id))
     }
@@ -252,15 +262,15 @@ class AssistanceController(
         clock: EvakaClock,
     ): List<AssistanceActionOption> {
         return db.connect { dbc ->
-                dbc.read {
+                dbc.read { tx ->
                     accessControl.requirePermissionFor(
-                        it,
+                        tx,
                         user,
                         clock,
                         Action.Global.READ_ASSISTANCE_ACTION_OPTIONS,
                     )
+                    tx.getAssistanceActionOptions()
                 }
-                assistanceActionService.getAssistanceActionOptions(dbc)
             }
             .also { Audit.AssistanceActionOptionsRead.log() }
     }
@@ -597,5 +607,25 @@ class AssistanceController(
             .also { deletedId ->
                 deletedId?.let { Audit.OtherAssistanceMeasureDelete.log(targetId = AuditId(it)) }
             }
+    }
+
+    private fun validateActions(
+        data: AssistanceActionRequest,
+        options: List<AssistanceActionOption>,
+    ) {
+        data.actions.forEach { action ->
+            val option =
+                options.find { it.value == action }
+                    ?: throw BadRequest(
+                        "Action $action is not a recognized option, all options: ${options.map { it.value }}"
+                    )
+
+            if (option.validFrom != null && data.startDate < option.validFrom) {
+                throw BadRequest("Action $action cannot be used before ${option.validFrom}")
+            }
+            if (option.validTo != null && data.endDate > option.validTo) {
+                throw BadRequest("Action $action cannot be used after ${option.validTo}")
+            }
+        }
     }
 }
